pyproject.toml
[tool.poetry]
name = "chatgpt"
version = "0.1.0"
description = ""
authors = ["Santiago-Balcero <santiagobalcero@gmail.com>"]
readme = "README.md"
[tool.poetry.scripts]
chatgpt = "chatgpt.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.7.0"}
openai = "^0.27.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
